+++ CHANGES.md
@@ -8,6 +8,8 @@
 
 * Fix steraming response handling for http/2
 
+* Better sleep on error support
+
 
 ## 0.4.6 (2018-03-10)